Create loader entrypoint binary
